Refactoring the directory structure.
