+++ estimation/unidimensional/dichotomous/partial_integrals.py
@@ -1,4 +1,31 @@
 import numpy as np
+from scipy.special import expit
+
+
+def _compute_partial_integral(theta, difficulty, discrimination, 
+                              responses, invalid_response_mask):
+    """ Computes the partial integral for an IRT item
+
+    The return array is 2D with dimensions
+        (person x theta)
+
+    Args:
+        theta: (1d array) Quadrature locations
+        difficulty: (float) Item difficulty parameter
+        discrimination: (float) Item discrimination parameter
+        responses: (1d array) Correct (1) or Incorrect (0)
+        invalid_response_mask: (1d array) mask where data is missing
+
+    Returns:
+        the_output: (2d array) is modified in-place
+    """                               
+    probability = expit(np.outer([-1, 1], 
+                        discrimination * (theta - difficulty)))
+    
+    # Mask out missing responses
+    temp_output = probability[responses, :]
+    temp_output[invalid_response_mask] = 1
+    return temp_output
 
 
 def _compute_partial_integral_3pl(theta, difficulty, discrimination, guessing, the_sign):
