--- src/three_pl/__init__.py
@@ -1,2 +1,0 @@
-from .three_pl_mml import threepl_mml
-from .ability_three_pl import ability_3pl_eap, ability_3pl_map, ability_3pl_mle
