+++ synthetic/__init__.py
@@ -1,0 +1,2 @@
+from .dichotomous import *
+from .polytomous import *
